Add support for namespace separators
This adds the ability to specify the namespace separator, to allow for
dry-container's configurable namespace separators. It performs some
introspection on the container to determine what the namespace separator
should be.

It's not hard-coded to dry-container since you _can_ pass in a
`:namespace_separator` option into the Builder that is then passed along
the chain. This value will actually override the value that is detected
from the container.

There are two main things that I don't like about this approach:

* There isn't a great interface for querying a dry-container for its
  namespace separator. Thus, we have to dig into the container's
  `#config` property to get at it. That feels subpar to me because we're
  "defining" a container as something with a `#config` that has a
  specific attribute instead of asking the container directly.
* Since the namespace separator has to be specified up front, we have to
  pass it along several layers that don't care about it. It becomes
  cruft in the initializers. One solution to this would be to build a
  configuration object that gets passed through the layers, but that
  still doesn't seem great. Any ideas?