Raise exception when injecting into modules
To me it happened that I accidentally injected my dependencies into a
namespace module instead of into a class, like this:

  module Operations
    module Articles
      include Import["repositories.articles"]

      class Create
        # ...
      end
    end
  end

When I did that, I got a NoMethodError for NilClass somewhere down the
line. This led me to believe that I've somehow incorrectly passed
parameters to `Import[]`, and it took me a while to realize what was the
error.

The error occurred because the downstream code assumes the #initialize
method will be defined, which is not the case for modules. To improve
the developer experience, we detect that we're attempting to inject into
a module and raise an explicit exception.